Rename SkipEmptyContractResolver
